adds cosmetic changes 1. modular controllers 2. modular models (models will get data from db or external apis (like gdax or cryptopanic)) 3. helpers contains business logic for gdax or cryptopanic 4. config for configurable parameters used by express server
