+++ server/controllers/index.js
@@ -1,81 +1,13 @@
-const models = require('../models');
-const gdax = require('../../helpers/gdax.js');
-const db = require('../db/index.js')
+const list = require('./list');
+const range = require('./range');
+const search = require('./search');
+const ticker = require('./ticker');
+const user = require('./user');
 
 module.exports = {
-  range: {
-    get: (req, res) => {
-      let { coin, dateStart, dateEnd, granularity } = req.query;
-      granularity /= 1000;
-      dateStart = Number(dateStart);
-      dateEnd = Number(dateEnd);
-      gdax.getTimeSeriesByRange(coin, dateStart, dateEnd, granularity)
-        .then((series) => {
-          res.json(series);
-        })
-        .catch((err) => {
-          console.error(err);
-          res.sendStatus(404);
-        });
-    },
-  },
-  ticker: {
-    get: (req, res) => {
-      const { coin } = req.query;
-      gdax.getTickerData(coin)
-        .then((ticker) => {
-          res.json(ticker);
-        })
-        .catch((err) => {
-          console.error(err);
-          res.sendStatus(404);
-        });
-    },
-  },
-  search: {
-    get: (req, res) => {
-      let { currency } = req.query; // get currency from req.query
-
-      models.search.get(currency)
-        .then((data) => {
-          res.status(200).send(data);
-        })
-        .catch((err) => {
-          res.status(404).send();
-          throw err;
-        });
-    },
-  },
-
-  list: {
-    get: (req, res) => {
-      models.list.get()
-        .then((data) => {
-          res.status(200).send(data);
-        })
-        .catch((err) => {
-          res.status(404).send();
-          throw err;
-        });
-    },
-    post: (req, res) => {
-
-    },
-  },
-
-  user: {
-    get: (req, res) => {
-      const { username } = req.query;
-      db.User.findOne({ username })
-        .then((data) => {
-          res.status(200).send(data || db.User.create({ username }));
-        })
-    },
-    post: (req, res) => {
-      const { username, coin, quantity } = req.body;
-      db.updatePosition(username, coin, Number(quantity))
-        .then(res.sendStatus(201));
-    },
-  },
-
+  list,
+  range,
+  search,
+  ticker,
+  user,
 };
