+++ server/models/list.js
@@ -1,0 +1,10 @@
+const Promise = require('bluebird');
+
+module.exports = {
+  get: () => {
+    let dummyData = 'List';
+    return new Promise((resolve, reject) => {
+      resolve(dummyData);
+    });
+  },
+};
