+++ server/models/range.js
@@ -1,0 +1,6 @@
+const gdax = require('../helpers/gdax');
+
+module.exports = {
+  getTimeSeriesByRange: (coin, dateStart, dateEnd, granularity) =>
+    gdax.getTimeSeriesByRange(coin, dateStart, dateEnd, granularity),
+};
