+++ server/models/search.js
@@ -1,0 +1,5 @@
+const cryptopanic = require('../helpers/cryptopanic');
+
+module.exports = {
+  getCryptopanicNews: coin => cryptopanic.getCryptopanicNews(coin),
+};
