+++ server/models/ticker.js
@@ -1,0 +1,5 @@
+const gdax = require('../helpers/gdax');
+
+module.exports = {
+  getTickerData: coin => gdax.getTickerData(coin),
+};
