+++ server/models/user.js
@@ -1,0 +1,10 @@
+const Promise = require('bluebird');
+
+module.exports = {
+  get: () => {
+    let dummyData = 'User';
+    return new Promise((resolve, reject) => {
+      resolve(dummyData);
+    });
+  },
+};
